+++ snippy/config/config.py
@@ -111,6 +111,7 @@ class Config(object):
         cls.content_category = cls.source.category
         cls.content_data = cls.source.data
         cls.content_brief = cls.source.brief
+        cls.content_description = cls.source.description
         cls.content_groups = cls.source.groups
         cls.content_tags = cls.source.tags
         cls.content_links = cls.source.links
@@ -257,6 +258,7 @@ class Config(object):
         resource = collection.get_resource(cls.content_category, Config.utcnow())
         resource.data = cls.content_data
         resource.brief = cls.content_brief
+        resource.description = cls.content_description
         resource.groups = cls.content_groups
         resource.tags = cls.content_tags
         resource.links = cls.content_links
@@ -543,6 +545,7 @@ class Config(object):
         cls._logger.debug('configured content category: %s', cls.content_category)
         cls._logger.debug('configured content data: %s', cls.content_data)
         cls._logger.debug('configured content brief: %s', cls.content_brief)
+        cls._logger.debug('configured content description: %s', cls.content_description)
         cls._logger.debug('configured content groups: %s', cls.content_groups)
         cls._logger.debug('configured content tags: %s', cls.content_tags)
         cls._logger.debug('configured content links: %s', cls.content_links)