+++ tests/test_api_search_solution.py
@@ -45,7 +45,7 @@ class TestApiSearchSolution(object):  # pylint: disable=too-many-public-methods
 
         result_headers = {
             'content-type': 'application/vnd.api+json; charset=UTF-8',
-            'content-length': '5426'
+            'content-length': '5464'
         }
         result_json = {
             'meta': {
@@ -87,7 +87,7 @@ class TestApiSearchSolution(object):  # pylint: disable=too-many-public-methods
 
         result_headers = {
             'content-type': 'application/vnd.api+json; charset=UTF-8',
-            'content-length': '5425'
+            'content-length': '5463'
         }
         result_json = {
             'meta': {
@@ -198,7 +198,7 @@ class TestApiSearchSolution(object):  # pylint: disable=too-many-public-methods
 
         result_headers = {
             'content-type': 'application/vnd.api+json; charset=UTF-8',
-            'content-length': '7093'
+            'content-length': '7131'
         }
         result_json = {
             'meta': {
@@ -238,7 +238,7 @@ class TestApiSearchSolution(object):  # pylint: disable=too-many-public-methods
 
         result_headers = {
             'content-type': 'application/vnd.api+json; charset=UTF-8',
-            'content-length': '7635'
+            'content-length': '7673'
         }
         result_json = {
             'meta': {
@@ -425,7 +425,7 @@ class TestApiSearchSolution(object):  # pylint: disable=too-many-public-methods
 
         result_headers = {
             'content-type': 'application/vnd.api+json; charset=UTF-8',
-            'content-length': '2569'
+            'content-length': '2588'
         }
         result_json = {
             'meta': {
@@ -488,7 +488,7 @@ class TestApiSearchSolution(object):  # pylint: disable=too-many-public-methods
 
         result_headers = {
             'content-type': 'application/vnd.api+json; charset=UTF-8',
-            'content-length': '5426'
+            'content-length': '5464'
         }
         result_json = {
             'meta': {
@@ -527,7 +527,7 @@ class TestApiSearchSolution(object):  # pylint: disable=too-many-public-methods
 
         result_headers = {
             'content-type': 'application/vnd.api+json; charset=UTF-8',
-            'content-length': '3018'
+            'content-length': '3037'
         }
         result_json = {
             'meta': {
@@ -562,7 +562,7 @@ class TestApiSearchSolution(object):  # pylint: disable=too-many-public-methods
 
         result_headers = {
             'content-type': 'application/vnd.api+json; charset=UTF-8',
-            'content-length': '8508'
+            'content-length': '8578'
         }
         result_json = {
             'meta': {